// http3/src/app/shared/interfaces/user.model.ts
export interface User {
  _id: string;
  id: number;
  guid: string;
  isActive: boolean;
  balance: string;
  picture: string;
  age: number;
  eyeColor: string;
  name: {
    first: string,
    last: string
  };
  gender: string;
  company: string;
  email: string;
  phone: string;
  address: string;
  about: string;
  registered: string;
  latitude: number;
  longitude: number;
  greeting: string;
}